+++ src/connector.ts
@@ -65,10 +65,6 @@ async function readRuntimeConfigs() {
 
 async function connectToHost(hostname: string, port: number, connectSequence: ReplConnectSequence) {
   state.analytics().logEvent('REPL', 'Connecting').send();
-  void state.analytics().logPlausiblePageview('/connecting', {
-    projectType: connectSequence?.projectType,
-    cljsType: connectSequence?.cljsType,
-  });
 
   if (nClient) {
     nClient['silent'] = true;
@@ -93,7 +89,7 @@ async function connectToHost(hostname: string, port: number, connectSequence: Re
         }
         // TODO: Figure out why the program bails out after here.
         // For now, we just clean up the connection state (even if we fail to return)
-        return cleanUpAfterError(connectSequence, e);
+        return cleanUpAfterError(e);
       },
     });
     nClient.addOnCloseHandler((c) => {
@@ -111,11 +107,7 @@ async function connectToHost(hostname: string, port: number, connectSequence: Re
     util.setConnectingState(false);
     util.setConnectedState(true);
     state.analytics().logEvent('REPL', 'ConnectedCLJ').send();
-    void state.analytics().logPlausiblePageview('/connected-clj', {
-      projectType: connectSequence?.projectType,
-      cljsType: connectSequence?.cljsType,
-    });
-
+    void state.analytics().storeFact('connected-clj-repl');
     setStateValue('clj', cljSession);
     setStateValue('cljc', cljSession);
     status.update();
@@ -170,26 +162,17 @@ async function connectToHost(hostname: string, port: number, connectSequence: Re
             isBuiltinType ? (connectSequence.cljsType as string) : 'Custom'
           )
           .send();
-        void state
-          .analytics()
-          .logPlausiblePageview(cljsSession ? '/connected-cljs' : 'connecting-cljs-failed', {
-            projectType: connectSequence?.projectType,
-            cljsType: connectSequence?.cljsType,
-          });
+        void state.analytics().storeFact('connected-cljs-repl');
       }
       if (cljsSession) {
         setUpCljsRepl(cljsSession, cljsBuild);
       }
     } catch (e) {
-      void state.analytics().logPlausiblePageview('connecting-cljs-failed', {
-        projectType: connectSequence?.projectType,
-        cljsType: connectSequence?.cljsType,
-      });
       outputWindow.appendLine('; Error while connecting cljs REPL: ' + e);
     }
     status.update();
   } catch (e) {
-    return cleanUpAfterError(connectSequence, e);
+    return cleanUpAfterError(e);
   }
 
   void liveShareSupport.didConnectRepl(port);
@@ -199,15 +182,11 @@ async function connectToHost(hostname: string, port: number, connectSequence: Re
   return true;
 }
 
-function cleanUpAfterError(connectSequence: ReplConnectSequence, e: any) {
+function cleanUpAfterError(e: any) {
   util.setConnectingState(false);
   util.setConnectedState(false);
   outputWindow.appendLine('; Failed connecting.');
   state.analytics().logEvent('REPL', 'FailedConnectingCLJ').send();
-  void state.analytics().logPlausiblePageview('/connecting-clj-failed', {
-    projectType: connectSequence?.projectType,
-    cljsType: connectSequence?.cljsType,
-  });
   console.error('Failed connecting:', e);
   status.update();
   return false;
@@ -640,19 +619,7 @@ export async function connect(
 
   state.analytics().logEvent('REPL', 'ConnectInitiated', isAutoConnect ? 'auto' : 'manual');
   state.analytics().logEvent('REPL', 'ConnectInitiated', cljsTypeName).send();
-  void state.analytics().logPlausiblePageview('/connect-initiated', {
-    isAutoConnect,
-    projectType: connectSequence?.projectType,
-    cljsType: connectSequence?.cljsType,
-    autoSelectForConnect: connectSequence?.autoSelectForConnect,
-    autoSelectForJackIn: connectSequence?.autoSelectForJackIn,
-    hasAfterCLJReplJackInCode: !!connectSequence?.afterCLJReplJackInCode,
-    hasCustomJackInCommandLine: !!connectSequence?.customJackInCommandLine,
-    hasJackInEnv: !!connectSequence?.jackInEnv,
-  });
-
   const portFile = projectTypes.nreplPortFileUri(connectSequence);
-
   void state.extensionContext.workspaceState.update('selectedCljsTypeName', cljsTypeName);
   void state.extensionContext.workspaceState.update('selectedConnectSequence', connectSequence);
 
@@ -719,15 +686,7 @@ async function standaloneConnect(
       .analytics()
       .logEvent('REPL', 'StandaloneConnect', `${connectSequence.name} + ${cljsTypeName}`)
       .send();
-    void state.analytics().logPlausiblePageview('/standalone-connect', {
-      projectType: connectSequence?.projectType,
-      cljsType: connectSequence?.cljsType,
-      autoSelectForConnect: connectSequence?.autoSelectForConnect,
-      autoSelectForJackIn: connectSequence?.autoSelectForJackIn,
-      hasAfterCLJReplJackInCode: !!connectSequence?.afterCLJReplJackInCode,
-      hasCustomJackInCommandLine: !!connectSequence?.customJackInCommandLine,
-      hasJackInEnv: !!connectSequence?.jackInEnv,
-    });
+    void state.analytics().storeFact('connect-initiated', 'standalone-connect');
     return connect(connectSequence, getConfig().autoSelectNReplPortFromPortFile, hostname, port);
   } else {
     outputWindow.appendLine('; Aborting connect, error determining connect sequence.');
@@ -856,9 +815,6 @@ export default {
     const cljsTypeName: string = state.extensionContext.workspaceState.get('selectedCljsTypeName'),
       cljTypeName: string = state.extensionContext.workspaceState.get('selectedCljTypeName');
     state.analytics().logEvent('REPL', 'switchCljsBuild', cljsTypeName).send();
-    void state.analytics().logPlausiblePageview('/switch-cljs-build', {
-      cljsType: cljsTypeName,
-    });
     const [session, build] = await makeCljsSessionClone(
       cljSession,
       translatedReplType,
@@ -866,10 +822,6 @@ export default {
     );
     if (session) {
       setUpCljsRepl(session, build);
-    } else {
-      void state.analytics().logPlausiblePageview('/switch-cljs-build-failed', {
-        cljsType: cljsTypeName,
-      });
     }
     status.update();
   },