+++ CHANGELOG.md
@@ -4,11 +4,15 @@ Changes to Calva.
 
 ## [Unreleased]
 
+## [2.0.355] - 2023-05-04
+
+- Fix: [REPL connect fails if requiring clojure.main/repl-requires fails](https://github.com/BetterThanTomorrow/calva/issues/2178)
+- Calva development: Only log app start with plausible.io
+
 ## [2.0.354] - 2023-05-03
 
-- Fix: [cljfmt isn't found unless project is started](https://github.com/BetterThanTomorrow/calva/issues/2078)
 - Fix: [”Resolve macro as” menu buttons are unreadable](https://github.com/BetterThanTomorrow/calva/issues/2156)
-- Calva development: Add [Plausible](plausible.io) analytics, intended to replace Google Analytics
+- Calva development: Add [Plausible](https://plausible.io) analytics, intended to replace Google Analytics
 
 ## [2.0.353] - 2023-04-17